Reorganized the project, distinguished application and utils, added logging to file from all components
