+++ ptp/configuration/__init__.py
@@ -1,25 +1,10 @@
-from .app_state import AppState
-from .component_factory import ComponentFactory
 from .config_interface import ConfigInterface
 from .config_registry import ConfigRegistry
 from .configuration_error import ConfigurationError
-from .globals_facade import GlobalsFacade
-from .key_mappings_facade import KeyMappingsFacade
-from .pipeline_manager import PipelineManager
-from .problem_manager import ProblemManager
-from .sampler_factory import SamplerFactory
-from .singleton import SingletonMetaClass
 #from configs_parsing import load_default_configuration_file
 
 __all__ = [
-    'AppState',
-    'ComponentFactory',
     'ConfigInterface',
     'ConfigRegistry',
     'ConfigurationError',
-    'GlobalsFacade',
-    'PipelineManager',
-    'ProblemManager',
-    'SamplerFactory',
-    'SingletonMetaClass',
     ]
