+++ ptp/utils/__init__.py
@@ -1,8 +1,16 @@
+from .app_state import AppState
+from .globals_facade import GlobalsFacade
+from .key_mappings_facade import KeyMappingsFacade
+from .singleton import SingletonMetaClass
 from .statistics_aggregator import StatisticsAggregator
 from .statistics_collector import StatisticsCollector
 
 
 __all__ = [
+    'AppState',
+    'GlobalsFacade',
+    'KeyMappingsFacade',
+    'SingletonMetaClass',
     'StatisticsAggregator',
     'StatisticsCollector',    
     ]
